Improve UI and responsiveness of technology section
This commit introduces several enhancements to the technology section of the website. Firstly, an import statement for `ArrowRight` from `lucide-preact` has been added to improve navigation cues. The console log for debugging technologies has been commented out to clean up the console output in production.

The link to the technologies page has been updated with additional styling and an arrow icon to make it more visually appealing and informative. The text has also been changed to "Технологии которые я использую" to better reflect the content's purpose.

In the carousel component, the background color changes on hover to improve user interaction feedback. The structure of the carousel items has been modified to include the link encompassing the image and title, making the entire area clickable and enhancing usability.

Significant improvements have been made to the carousel's responsiveness. The number of slides displayed now adjusts based on the window width, ensuring a better viewing experience across different device sizes. This adjustment is dynamically applied on window resize, making the carousel fully responsive.

These changes collectively enhance the user interface and interaction with the technology section, making it more engaging and accessible across various devices.